requirements.txt
# Core dependencies
pyyaml>=6.0
playwright>=1.40.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API dependencies (for web UI)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Firebase (if using)
firebase-admin>=6.2.0

# Other dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Vector Database (for long-term memory)
qdrant-client>=1.7.0
sentence-transformers>=2.2.0  # For local embeddings

# ============================================
# Telegram Bot Dependencies
# ============================================
python-telegram-bot>=20.0
openai>=1.0.0          # For /gpt commands

# ============================================
# Optional Third-party Integrations
# ============================================
# Install only the integrations you need:

# AI Integrations (optional)
# anthropic>=0.7.0       # For Anthropic/Claude integration (coming soon)
# google-generativeai    # For Google Gemini integration (coming soon)
